Rename extension_policy_builder.rs -> extension_policy.rs
